ci: enable ThinLTO for PGO build
since we've switched to fuchsia clang, the cost of ThinLTO construction has been greatly reduced, and we don't really need to enable ThinLTO for instrumented LLVM.
ref: https://github.com/llvm/llvm-project/issues/59506#issuecomment-1351178416

This is only for PGO build. For users who only build mpv occasionally and really care about the 15 minutes of extra build time, there is the option of not using PGO or using GCC toolchain, which will still be supported for a long time to come.